move the data view components to a subfolder
