+++ apps/newsletters-ui/src/app/components/views/DraftDetailView.tsx
