+++ apps/newsletters-ui/src/app/components/views/DraftListView.tsx
